taint/test/mochitest/test_dom_parser.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Taint DOM Parser propagation</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

        function todo_tainted(str) {
            todo(str.taint.length > 0, "Check tainted: " + str);
        }

        function check_tainted(str) {
            ok(str.taint.length > 0, "Check tainted: " + str);
        }

        function check_untainted(str) {
            ok(str.taint.length == 0, "Check untainted: " + str);
        }

        function testDomParser(str) {
            const parser = new DOMParser();
            check_tainted(str);

            const doc1 = parser.parseFromString(str, "text/html");

            check_tainted(doc1.body.firstChild.outerHTML);
            check_tainted(doc1.body.firstChild.textContent);
            check_tainted(doc1.body.innerHTML);
            check_tainted(doc1.body.outerHTML);
            check_tainted(doc1.body.textContent);
            return doc1;
        }

        function startTests() {

            const tainted_str = String.tainted("This is tainted text");

            doc1 = testDomParser("<p>" + tainted_str + "</p>");
            doc1 = testDomParser("<div>" + tainted_str + "</div>");
            doc1 = testDomParser("<div attr='" + String.tainted("value") + "'>" + tainted_str + "</div>");
            check_tainted(doc1.body.firstChild.getAttribute("attr"));

            const parser = new DOMParser();

            const tainted_comment_str = "<div></div><!-- " + tainted_str + " -->";
            const doc_comment = parser.parseFromString(tainted_comment_str, "text/html");
            check_tainted(doc_comment.body.childNodes[1].textContent);

            const tainted_foster_str = "<table>" + tainted_str;
            const doc_fostered = parser.parseFromString(tainted_foster_str, "text/html");
            check_tainted(doc_fostered.body.childNodes[0].textContent);

            const tainted_xml_str = "<warning>" + tainted_str + "</warning>";
            const doc2 = parser.parseFromString(tainted_xml_str, "application/xml");

            // This test is currently failing as the XML parser is not taint-aware
            // I have added some notes here in case this should be fixed in the future
            // DOMParser::ParseFromString --> creates and input stream and calls ParseFromStream
            //
            // DOMParser::ParseFromStream loads a new Document via an inputStream, which starts XML parsing,
            // using the XML parser, which uses xmlparse. On processing text, a callback tree is triggered:
            //
            // parser/expat/lib/xmlparse.c characterDataHandler --Callback--> nsExpatDriver::HandleCharacterData --Callback--> nsXMLContentSink::HandleCharacterData
            //
            // The whole chain above (xmlparse to nsXMLContentSink) is not taint aware.
            // To make this test work, all of the above components need to made taint-aware, including the xmlparse library.
            todo_tainted(doc2.documentElement.textContent);

            SimpleTest.finish();
        }
    </script>
  </head>
  <body onload="startTests()">
  </body>
</html>
